requirements.txt
openpyxl
google.generativeai
tk
markdown2
vosk
pyttsx3
streamlit
sounddevice
Flask
pywebview